Split knit merge.rs into run/compat/report modules
Break the 1,313-line src/commands/merge.rs into run (plan and apply the merge
run, continue/abort/rollback, finalize on the bundle), compat (build a
compatibility bundle from source bundles), and report (merge status/show and
pushing a recorded run). merge/mod.rs keeps the merge_command/create_compat
entrypoints, the run/step data types, and shared locking and git helpers. No
behavior change; tests and clippy unchanged.


Knit-Group: kg_20260531_90509a
Knit-Bundle: split-oversized-files-round-4
